Minor corrections of mismatch in use of 'int' and 'long'

+++ SRC/mon2_posdiv.c
@@ -303,7 +303,7 @@ void  OwnInit(int argc, char *argv[])
   {
     IntYZ     [iY] = (double*) malloc(nBinsDiv * sizeof(double));
     IntYZError[iY] = (double*) malloc(nBinsDiv * sizeof(double));
-    nTrajYZ   [iY] = (long*)   malloc(nBinsDiv * sizeof(int));
+    nTrajYZ   [iY] = (long*)   malloc(nBinsDiv * sizeof(long));
   }
 
   return;
